requirements.txt
python-dotenv
pydantic
requests
fastapi
uvicorn
openai